+++ testScripts/testPolling.js
@@ -2,10 +2,10 @@ require('dotenv').config()
 const LokiJS = require('../index')
 
 const loki = new LokiJS({
-    // Requests: {
-    //     url: 'https://valhalla.torus.topl.co:9585/',
-    //     apiKey: process.env.VALHALLA_KEY
-    // },
+    Requests: {
+        url: 'https://valhalla.torus.topl.co',
+        apiKey: process.env.VALHALLA_KEY
+    },
     KeyManager: {
         password: 'genesis',
         keyPath: './keystore/itGuy.json'