+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.<br/>
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).<br/>
+## 3.0.2 - 2020.04.10
+### Added
+* Stress test scripts to query Torus
+### Changed
+* Valhalla chain provider URL to `https://valhalla.torus.topl.co'
 ## 3.0.1 - 2020.04.09
 ### Added
 * Primary `Loki` class incorporating `KeyManager` and `Requests`